requirements.txt
flask
python-dotenv
supabase
sqlalchemy
psycopg2-binary